requirements.txt
anthropic>=0.40.0
openai>=1.50.0
python-dotenv>=1.0.0
httpx>=0.27.0
pydantic>=2.0.0
fastapi
uvicorn
langchain
langgraph
langchain-community
langchain-openai
langchain-chroma
langchain-text-splitters
chromadb
faiss-cpu
pymupdf
tiktoken
rank_bm25
numpy
